+++ src/MigrationTools/CommandLine/Execute/ExecuteCommand.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MigrationTools.CommandLine.Execute
+{
+    class ExecuteCommand
+    {
+    }
+}
